quinn-proto: move ShouldTransmit into streams module
